split registry to separate classes
