+++ osquery/registry/plugin.cpp
@@ -1,0 +1,51 @@
+/**
+ *  Copyright (c) 2018-present, Facebook, Inc.
+ *  All rights reserved.
+ *
+ *  This source code is licensed under both the Apache 2.0 license (found in the
+ *  LICENSE file in the root directory of this source tree) and the GPLv2 (found
+ *  in the COPYING file in the root directory of this source tree).
+ *  You may select, at your option, one of the above-listed licenses.
+ */
+
+#include <osquery/plugin.h>
+
+#include <boost/property_tree/json_parser.hpp>
+#include <boost/property_tree/ptree.hpp>
+
+namespace osquery {
+namespace pt = boost::property_tree;
+void Plugin::setName(const std::string& name) {
+  if (!name_.empty() && name != name_) {
+    std::string error = "Cannot rename plugin " + name_ + " to " + name;
+    throw std::runtime_error(error);
+  }
+
+  name_ = name;
+}
+
+void Plugin::getResponse(const std::string& key,
+                         const PluginResponse& response,
+                         boost::property_tree::ptree& tree) {
+  for (const auto& item : response) {
+    boost::property_tree::ptree child;
+    for (const auto& item_detail : item) {
+      child.put(item_detail.first, item_detail.second);
+    }
+    tree.add_child(key, child);
+  }
+}
+
+void Plugin::setResponse(const std::string& key,
+                         const boost::property_tree::ptree& tree,
+                         PluginResponse& response) {
+  std::ostringstream output;
+  try {
+    pt::write_json(output, tree, false);
+  } catch (const pt::json_parser::json_parser_error& /* e */) {
+    // The plugin response could not be serialized.
+  }
+  response.push_back({{key, output.str()}});
+}
+
+} // namespace osquery
